# exports/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "Events API",
    "description": "API for creating and managing events",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ProfileDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "Full name of the user",
            "example": "John Doe"
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the user in Nigerian format",
            "example": "+2348012345678"
          }
        },
        "required": [
          "fullName",
          "phone"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address of the user",
            "example": "oluwasheges@gmail.com"
          },
          "password": {
            "type": "string",
            "description": "Password for the user account",
            "example": "P@ssw0rd123"
          },
          "role": {
            "type": "string",
            "description": "Role of the user (creator or eventee)",
            "example": "creator",
            "enum": [
              "creator",
              "eventee"
            ]
          },
          "profile": {
            "description": "Profile information of the user",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProfileDto"
              }
            ]
          }
        },
        "required": [
          "email",
          "password",
          "role",
          "profile"
        ]
      },
      "UserProfileDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Oluwasegun Adedeji"
          },
          "phone": {
            "type": "string",
            "example": "08132638235"
          }
        },
        "required": [
          "fullName",
          "phone"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "oluwasheges@gmail.com"
          },
          "role": {
            "type": "string",
            "example": "creator"
          },
          "profile": {
            "$ref": "#/components/schemas/UserProfileDto"
          },
          "isVerified": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "email",
          "role",
          "profile",
          "isVerified"
        ]
      },
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "User registered successfully"
          },
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "required": [
          "message",
          "accessToken",
          "user"
        ]
      },
...
    }
  }
}