# src/users/dto/create-user.dto.ts
// export class CreateUserDto {
//   email: string;
//   password: string
//   role: 'creator' | 'eventee';
//   profile: {
//     fullName: string;
//     phone?: string;
//     avatar?: string;
//   };
// }